Trait justjson::parser::ParseDelegate

pub trait ParseDelegate<'a> {
    type Value;
    type Object;
    type Array;
    type Key;
    type Error;

    // Required methods
    fn null(&mut self) -> Result<Self::Value, Self::Error>;
    fn boolean(&mut self, value: bool) -> Result<Self::Value, Self::Error>;
    fn number(
        &mut self,
        value: JsonNumber<'a>
    ) -> Result<Self::Value, Self::Error>;
    fn string(
        &mut self,
        value: JsonString<'a>
    ) -> Result<Self::Value, Self::Error>;
    fn begin_object(&mut self) -> Result<Self::Object, Self::Error>;
    fn object_key(
        &mut self,
        object: &mut Self::Object,
        key: JsonString<'a>
    ) -> Result<Self::Key, Self::Error>;
    fn object_value(
        &mut self,
        object: &mut Self::Object,
        key: Self::Key,
        value: Self::Value
    ) -> Result<(), Self::Error>;
    fn object_is_empty(&self, object: &Self::Object) -> bool;
    fn end_object(
        &mut self,
        object: Self::Object
    ) -> Result<Self::Value, Self::Error>;
    fn begin_array(&mut self) -> Result<Self::Array, Self::Error>;
    fn array_value(
        &mut self,
        array: &mut Self::Array,
        value: Self::Value
    ) -> Result<(), Self::Error>;
    fn array_is_empty(&self, array: &Self::Array) -> bool;
    fn end_array(
        &mut self,
        array: Self::Array
    ) -> Result<Self::Value, Self::Error>;
    fn kind_of(&self, value: &Self::Value) -> JsonKind;

}

A delegate for a Parser.

This type has its functions invoked while a parser is parsing a JSON payload.

Required Associated Types

type Value

The type that can represent all parsed JSON value types.

type Object

The type that is used to represent a JSON object.

type Array

The type that is used to represent a JSON array.

type Key

The type that is used to represent the key of a field in a JSON object.

type Error

The error type for this delegate.

Required Methods

fn null(&mut self) -> Result<Self::Value, Self::Error>

Returns the value representation of null.

fn boolean(&mut self, value: bool) -> Result<Self::Value, Self::Error>

Returns the value representation of a boolean.

fn number(&mut self, value: JsonNumber<'a>) -> Result<Self::Value, Self::Error>

Returns the value representation of a JsonNumber.

fn string(&mut self, value: JsonString<'a>) -> Result<Self::Value, Self::Error>

Returns the value representation of a JsonString.

fn begin_object(&mut self) -> Result<Self::Object, Self::Error>

Returns an empty object.

fn object_key( &mut self, object: &mut Self::Object, key: JsonString<'a> ) -> Result<Self::Key, Self::Error>

Processes the key for a new value in an object. Returns the key representation of the JsonString.

fn object_value( &mut self, object: &mut Self::Object, key: Self::Key, value: Self::Value ) -> Result<(), Self::Error>

Adds a new key-value pair to an object.

fn object_is_empty(&self, object: &Self::Object) -> bool

Returns true if the object passed is empty.

fn end_object( &mut self, object: Self::Object ) -> Result<Self::Value, Self::Error>

Returns the value representation of the object passed.

fn begin_array(&mut self) -> Result<Self::Array, Self::Error>

Returns an empty array.

fn array_value( &mut self, array: &mut Self::Array, value: Self::Value ) -> Result<(), Self::Error>

Adds a new value to an array.

fn array_is_empty(&self, array: &Self::Array) -> bool

Returns true if the array passed is empty.

fn end_array(&mut self, array: Self::Array) -> Result<Self::Value, Self::Error>

Returns the value representation of the array passed.

fn kind_of(&self, value: &Self::Value) -> JsonKind

Returns the JsonKind of value.