Struct justjson::JsonString

pub struct JsonString<'a> { /* private fields */ }

A JSON string.

Internally, this type may be represented by either an already escaped JSON string or an unescaped string.

This type implements Eq and PartialEq such that the escapes are always handled correctly. Consider the following examples:

use justjson::JsonString;

let json = JsonString::from_json(r#""Check out this cat: \"\ud83d\ude39\"""#).unwrap();
assert_eq!(json, "Check out this cat: \"😹\"");
let alternate_form =
    JsonString::from_json("\"Check out this cat: \\\"\u{1f639}\\\"\"").unwrap();
assert_eq!(json, alternate_form)

When the underlying representation is an encoded JSON string, the JsonStringInfo is stored from the parsing operation. This allows the comparison implementations to perform quick length checks on the decoded length to avoid comparing the bytes of strings that decode to different lengths.

Additionally, this type implements Ord, PartialOrd, and Hash in ways that are consistent regardless of whether the string has escape sequences or not. If the underlying representation does not need extra processing, the built-in implementations for &str are always used.

Implementations

impl<'a> JsonString<'a>

pub fn from_json(json: &'a str) -> Result<Self, Error>

Parses json, expecting a single string value.

Errors

Returns ErrorKind::ExpectedString if a non-string value is encountered.

pub fn decode_if_needed(&self) -> AnyStr<'a>

Available on crate feature alloc only.

Returns the contained string after decoding any escape sequences, if needed. If there are no escape sequences, the contents of the string are borrowed.

This type keeps track of whether the underlying string contains escape sequences, so this function is nearly instananeous when there are no escape sequences.

pub fn escape_if_needed(&self) -> AnyStr<'a>

Available on crate feature alloc only.

Returns the contained string after escaping any characters, if needed. If the string is already internally represented as a JSON string, this function returns its source via borrowing and no additional processing.

pub fn decoded(&self) -> Decoded<'_>

Returns the string after decoding any JSON escape sequences.

pub fn as_json(&self) -> AsJson<'_>

Returns the string after decoding any JSON escape sequences.

pub fn as_json_str(&self) -> Option<&AnyStr<'a>>

Returns a reference to the underlying storage of this string, if it’s already encoded for use in a JSON string. This does not include the surrounding quotation marks.

If None is returned, the string must be decoded using as_json().

pub fn len(&self) -> usize

Returns the length of this string when encoded as JSON.

pub fn is_empty(&self) -> bool

Returns true if this string is empty.

pub fn decoded_len(&self) -> usize

Returns the length of this string when decoded.

pub fn as_str(&self) -> Option<&str>

Returns a reference to the contents of this value if the contained string does not have any escape sequences that must be decoded.