Trait justjson::parser::ParseDelegate

pub trait ParseDelegate<'a> {
    type Value;
    type Object;
    type Array;
    type Key;

    fn null(&mut self) -> Self::Value;
    fn boolean(&mut self, value: bool) -> Self::Value;
    fn number(&mut self, value: JsonNumber<'a>) -> Self::Value;
    fn string(&mut self, value: JsonString<'a>) -> Self::Value;
    fn begin_object(&mut self) -> Self::Object;
    fn object_key(
        &mut self,
        object: &mut Self::Object,
        key: JsonString<'a>
    ) -> Self::Key;
    fn object_value(
        &mut self,
        object: &mut Self::Object,
        key: Self::Key,
        value: Self::Value
    );
    fn object_is_empty(&self, object: &Self::Object) -> bool;
    fn end_object(&mut self, object: Self::Object) -> Self::Value;
    fn begin_array(&mut self) -> Self::Array;
    fn array_value(&mut self, array: &mut Self::Array, value: Self::Value);
    fn array_is_empty(&self, array: &Self::Array) -> bool;
    fn end_array(&mut self, array: Self::Array) -> Self::Value;
    fn kind_of(&self, value: &Self::Value) -> JsonKind;

}

A delegate for a Parser.

This type has its functions invoked while a parser is parsing a JSON payload.

Required Associated Types

type Value

The type that can represent all parsed JSON value types.

type Object

The type that is used to represent a JSON object.

type Array

The type that is used to represent a JSON array.

type Key

The type that is used to represent the key of a field in a JSON object.

Required Methods

fn null(&mut self) -> Self::Value

Returns the value representation of null.

fn boolean(&mut self, value: bool) -> Self::Value

Returns the value representation of a boolean.

fn number(&mut self, value: JsonNumber<'a>) -> Self::Value

Returns the value representation of a JsonNumber.

fn string(&mut self, value: JsonString<'a>) -> Self::Value

Returns the value representation of a JsonString.

fn begin_object(&mut self) -> Self::Object

Returns an empty object.

fn object_key(
&mut self,
object: &mut Self::Object,
key: JsonString<'a>
) -> Self::Key

Processes the key for a new value in an object. Returns the key representation of the JsonString.

fn object_value(
&mut self,
object: &mut Self::Object,
key: Self::Key,
value: Self::Value
)

Adds a new key-value pair to an object.

fn object_is_empty(&self, object: &Self::Object) -> bool

Returns true if the object passed is empty.

fn end_object(&mut self, object: Self::Object) -> Self::Value

Returns the value representation of the object passed.

fn begin_array(&mut self) -> Self::Array

Returns an empty array.

fn array_value(&mut self, array: &mut Self::Array, value: Self::Value)

Adds a new value to an array.

fn array_is_empty(&self, array: &Self::Array) -> bool

Returns true if the array passed is empty.

fn end_array(&mut self, array: Self::Array) -> Self::Value

Returns the value representation of the array passed.

fn kind_of(&self, value: &Self::Value) -> JsonKind

Returns the JsonKind of value.